function/requirements.txt
azure-functions
DateTime==4.3
psycopg2-binary==2.8.5
python-http-client==3.2.7
pytz==2020.1
sendgrid==6.3.1
zope.interface==5.1.0
